requirements.txt
python-dotenv
requests

# For Testing
#flake8
#pylint
